+++ mm2src/common/common.rs
@@ -118,6 +118,7 @@ pub mod custom_futures;
 pub mod custom_iter;
 #[path = "executor/mod.rs"] pub mod executor;
 pub mod number_type_casting;
+pub mod password_policy;
 pub mod seri;
 #[path = "patterns/state_machine.rs"] pub mod state_machine;
 pub mod time_cache;