+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## {inc-release}
 
 **Features:**
+- `adex-cli` command line utility was introduced that supplies commands: `init`, `start`, `stop`, `status` [#1729](https://github.com/KomodoPlatform/atomicDEX-API/pull/1729)
 
 **Enhancements/Fixes:**
 - CI/CD workflow logics are improved [#1736](https://github.com/KomodoPlatform/atomicDEX-API/pull/1736)